+++ src/sections/mywords/mywords.css
@@ -22,3 +22,19 @@
   font-weight: 700;
   line-height: normal;
 }
+
+.test-relative {
+  /* position: relative; */
+  background: #ffecea;
+}
+
+.test-relative > img {
+  z-index: 1;
+  position: relative;
+  top: -9px;
+}
+
+.test-relative > div {
+  z-index: 2;
+  position: relative;
+}
